Add internal and user config
This commit creates an internal config object, with the ability
for the user to load and override default config settings.

The user config is set through the git config, either .gitconfig
or .git/config.  If the user config contains values that are set
in the internal config, the user config values will be used.