Remove GoFmt commit check module
The GoFmt.pm module was originally added as an example for adding
different kinds of checks to run through hooks.

Since it relies on an external program, gofmt, it should exist as
a dist on it's own, not within the core check modules; which are
meant to be agnostic to programming languages themselves.