Add commit checks IndentSpaces and IndentTabs
This commit adds the git commit check modules, IndentSpaces and
IndentTabs.  IndentSpaces checks the commit and warns if indent
spaces are found.  IndentTabs checks the commit and warns if indent
tabs are found.

Since the default commit check profile runs all commit checks by
default, an unchanged default profile will not allow the user to
add a commit since they'll get a warning for either tabs or spaces.
Both of these checks also overlap a little with the
MixedIndentTabsSpaces commit check.  A user may find two warnings
given if mixed indent tabs and spaces are found; one for the tab
indent and one for the mixed indent tab.